tests/wpt/tests/fetch/compression-dictionary/dictionary-clear-site-data-cache.tentative.https.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<meta name="timeout" content="long"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/compression-dictionary-util.js"></script>
</head>
<body>
<script>

compression_dictionary_promise_test(async (t) => {
  const dict = await (await fetch(kRegisterDictionaryPath)).text();
  // Wait until `available-dictionary` header is available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {}),
      kDefaultDictionaryHashBase64);
  // Clear site data.
  assert_equals(await clearSiteData(/*directive=*/'cache'), 'OK');
  // Check if `available-dictionary` header is not available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {max_retry: 0}),
      '"available-dictionary" header is not available');
}, 'Clear-Site-Data with "cache" directive must unregister dictionary');

</script>
</body>
